requirements.txt
av==8.0.3
Cython==0.29.22
imageio==2.8.0
imageio-ffmpeg==0.4.2
imgaug==0.4.0
ipdb==0.13.4
ipython==7.14.0
matplotlib==3.2.1
numpy==1.18.4
opencv-python-headless==4.5.3.56
pandas==1.0.5
parso==0.7.0
pdbpp==0.10.2
Pillow==7.1.2
pkg-resources==0.0.0
pycocotools==2.0.2
PyYAML==6.0
scikit-image==0.17.2
scikit-learn==0.24.2
scikit-video==1.1.11
scipy==1.4.1
torch==1.5.0
torchaudio==0.10.1
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.6.0
tornado==6.1
tqdm==4.58.0
urllib3==1.26.3